fix(PreServiceEvaluation): Improve type safety for updates
Introduces type safety for the `field` parameter in `handlePreServiceIndividualUpdate` by changing it from `string` to `keyof PreServiceIndividualEvaluation`. This change, along with the use of a switch statement, ensures that only valid fields are updated, preventing runtime errors and improving code robustness.

Also, ensures `raId` is set on `CriterioEvaluacion` before saving.